Move Scramble to own folder
