+++ client/src/components/Scramble/Score.js
